Did some prep work for the update to forum key
The forum key is formatted differently then a super user key, which is
what was being used before. A super user key was like localStorage
except it could be accessed by anyone if they were viewing a page that
the user was loaded on. However, a global forum key is one giant key
that everyone accesses. So the get and set functions all have to be
changed. So instead of using proboard's key "get" function. I changed
all of the code to use my get function. So instead of having to change
multiple instances, I'll just have to change the one function.